move lerp and clamp so I dont have duplicate code
