+++ src/seahawk_rov/seahawk_rov/__init__.py
@@ -33,3 +33,19 @@ from seahawk_rov.thrust_box_bme280 import ThrustBoxBME280
 from seahawk_rov.logic_tube_motor import LogicTubeMotor
 from seahawk_rov.logic_tube_servo import LogicTubeServo
 from seahawk_rov.thrust_box_servo import ThrustBoxServo
+
+
+# linear interpolation helper function
+def lerp(old_min:float, old_max:float, new_min:int, new_max:int, old_value:float):
+    '''linear interpolate helper function'''
+    old_range = old_max - old_min
+    new_range = new_max - new_min
+    new_value = (((old_value - old_min) * new_range) / old_range) + new_min
+    return new_value
+
+
+# clamp helper function
+# helps keep us from starting fires
+def clamp(num, minimum, maximum):
+  '''clamp helper function'''
+  return max(min(minimum, num), maximum)
